test(clicker): turmoil connection fidelity plus offline/turmoil reorg
Split the mesh harness into tests/offline (in-memory systems) and
tests/turmoil (virtual-network), and lift the rig into
src/testing/turmoil/. Then raise the rig's connection-layer fidelity so
the real mainnet failure modes from MAINNET_PROBLEMS 1-3 are reproducible
fast, and document it in MAINNET_PROBLEMS_8.md.

Rig:
- dial_with_deadline: bounded connect retries surface DialFailed instead
  of hanging the sim (covers dial timeout + partition-during-handshake).
- Nat/NoHairpin + public_ip/dialable + GhostHint: address model and dead
  ports, so same-public-IP dials fail and stale hints are dialed then
  dropped.
- discovery-driven dial: persistent accept_loop, length-prefixed hello
  handshake, register, higher-id-dials tie-break (plan_for) and
  redial_missing on a period, replacing link_app's synthetic
  PeerConnected flood.
- seed matrix over [7,8,9] for the retained-leave and diverged-repair
  cases; new connect tests (dial deadline, hairpin, cold-start stagger).

Rig incident: the first bring_up dropped its listener on return, so no
peer could ever reconnect and slice-3 convergence was non-deterministic;
a persistent accept loop fixes it.

Also enforce --all-targets --all-features on cargo tasks in
lele_enforce_config, add a telegram_notify helper test, and provision
telegram_bot_macros devenv files.

Turmoil 17/17 green; clicker 325 passed (1 pre-existing DISPLAY-only UI
test); clippy/fmt/lele_lint/bevy_lint/taxonomy green.

+++ clicker/src/testing/click_at_y.rs
@@ -7,8 +7,8 @@ use super::window_size;
 pub fn click_at_y(title: &str, fx: f64, y: i32) -> Result<(), String> {
     let wid = window_id::window_id(title)?;
     let (width, _) = window_size::window_size(&wid)?;
-    let x = (f64::from(width) * fx) as i32;
-    mouse_click_at::mouse_click_at(&wid, x, y)
+    let x = f64::from(width) * fx;
+    mouse_click_at::mouse_click_at(&wid, x, f64::from(y))
 }
 
 #[cfg(test)]
